// libs/enums/common.enum.ts
export enum Message {
  SOMETHING_WENT_WRONG = 'Something went wrong!',
  NO_DATA_FOUND = 'No data found!',
  CREATE_FAILED = 'Create failed!',
  UPDATE_FAILED = 'Update failed!',
  REMOVE_FAILED = 'Remove failed!',

  USED_NICK_PHONE = 'You are inserting already used nick or phone!',
  TOKEN_CREATION_FAILED = 'Token creation error!',
  NO_MEMBER_NICK = 'No member with that member nick!',
  BLOCKED_USER = 'You have been blocked!',
  WRONG_PASSWORD = 'Wrong password, try again!',
  NOT_AUTHENTICATED = 'Please login first!',
  TOKEN_NOT_EXIST = 'Bearer Token is not provided!',
  ONLY_SPECIFIC_ROLES_ALLOWED = 'Allowed only for members with specific roles!',
  NOT_ALLOWED_REQUEST = 'Not Allowed Request!',
  INSERT_ALL_INPUTS = 'Please provide all inputs',
}

export enum Direction {
  ASC = 'ASC',
  DESC = 'DESC',
}